+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Stages are now the central part; multiple project can be associated with the sam
 
 * added `attach` and `detatch` commands to manage project - stage association
 * dumped vhosts bind mounts, copy war instead
+* use image tags with timestamps
 
 * stool no longer adjust the current working directory
   * dumped cd (TODO: add this functionality to pommes)